Add Dockerfile to create a build image
The dockerfile creates an image based on ubuntu 20.04 that installs R,
some r-cran-* dependencies that we use, and additional build tools
(pandoc, make, g++ and gfortran, etc).

This image is then ready to build the Rmarkdown files into a static html
site. Dependencies that can are installed from ubuntu packages are in
`ubuntu-packages.list`