use github docker registry
Build and publish the docker image to the github registry,
and then, use the same image from the github registry when building the
gh-pages.

This was not possible previously since the github registry didn't
support public images, and github actions didn't support non-public
images (or it was cumbersome).

But now that changed:
https://github.blog/2020-09-01-introducing-github-container-registry/

Also with the announcement by Docker that they will remove inactive
images, Github registry seems like a better option.